+++ src/main/kotlin/study/domain/Person.kt
@@ -1,0 +1,3 @@
+package study.domain
+
+data class Person(val name: String, val company: String, val skills: Skills)
