+++ src/main/kotlin/study/domain/Skill.kt
@@ -1,3 +1,3 @@
-package study
+package study.domain
 
 data class Skill(val type: String, val description: String)
