+++ src/main/kotlin/study/domain/Skills.kt
@@ -1,0 +1,17 @@
+package study.domain
+
+class Skills {
+    private val list = mutableListOf<Skill>()
+
+    fun soft(description: String): Skills {
+        list.add(Skill("Soft", description))
+        return this
+    }
+
+    fun hard(description: String): Skills {
+        list.add(Skill("Hard", description))
+        return this
+    }
+
+    operator fun invoke(): List<Skill> = list.toList()
+}
