+++ src/main/kotlin/study/dsl/Introduce.kt
@@ -1,4 +1,7 @@
-package study
+package study.dsl
+
+import study.builder.PersonBuilder
+import study.domain.Person
 
 fun introduce(block: PersonBuilder.() -> Unit): Person {
     // also, apply, run, let, with
